requirements.txt
tensorflow[and-cuda]
PyYAML
matplotlib
IPython
scikit-image
opencv-python
pandas